[BPF] cleaning up maps is for both TC and XDP
Move it from tc package to bpf package with some code reorganization for
simplicity.

Also lock the clean up lock whenupdating XDP. It wasn't synced till now.
