+++ felix/bpf/ut/attach_test.go
@@ -74,7 +74,7 @@ func TestReattachPrograms(t *testing.T) {
 
 	// Start with a clean base state in case another test left something behind.
 	t.Log("Doing initial clean up")
-	tc.CleanUpMaps()
+	bpf.CleanUpMaps()
 	bpf.CleanAttachedProgDir()
 
 	startingJumpMaps := countJumpMaps()
@@ -141,7 +141,7 @@ func TestReattachPrograms(t *testing.T) {
 
 	// Clean up maps, but nothing should change
 	t.Log("Cleaning up, should remove the first map")
-	tc.CleanUpMaps()
+	bpf.CleanUpMaps()
 	Expect(countJumpMaps()).To(BeNumerically("==", startingJumpMaps+3), "unexpected number of jump maps")
 	Expect(countTCDirs()).To(BeNumerically("==", startingTCDirs+3), "unexpected number of TC dirs")
 	Expect(countHashFiles()).To(BeNumerically("==", startingHashFiles+3), "unexpected number of hash files")
@@ -155,7 +155,7 @@ func TestReattachPrograms(t *testing.T) {
 	Expect(err).NotTo(HaveOccurred())
 	err = tc.RemoveQdisc(vethName2)
 	Expect(err).NotTo(HaveOccurred())
-	tc.CleanUpMaps()
+	bpf.CleanUpMaps()
 	Expect(countJumpMaps()).To(BeNumerically("==", startingJumpMaps+1), "unexpected number of jump maps")
 	Expect(countTCDirs()).To(BeNumerically("==", startingTCDirs+1), "unexpected number of TC dirs")
 	Expect(countHashFiles()).To(BeNumerically("==", startingHashFiles+1), "unexpected number of hash files")
@@ -180,7 +180,7 @@ func TestReattachPrograms(t *testing.T) {
 	t.Log("Removing the XDP program and cleaning up its jump map, should return to base state")
 	err = ap3.DetachProgram()
 	Expect(err).NotTo(HaveOccurred())
-	tc.CleanUpMaps()
+	bpf.CleanUpMaps()
 	Expect(countJumpMaps()).To(BeNumerically("==", startingJumpMaps), "unexpected number of jump maps")
 	Expect(countTCDirs()).To(BeNumerically("==", startingTCDirs), "unexpected number of TC dirs")
 	Expect(countHashFiles()).To(BeNumerically("==", startingHashFiles), "unexpected number of hash files")
